Initial Commit
So far all we have is xlcall.h modified and converted using rust-bindgen into xlcall.rs, which was then modified to match the calling conventions of the Excel4 and Excel12 addins. Not yet tested in Excel (I do not yet have access to a machine running Windows).

Next steps, to test it and wrap up the XLOPER struct with something much more rust-friendly.